Refactored the interpreter.
Dictionary-like words (find, number, and ParseInstruction) now all
have a standard interface. They have these stack effects:

( str-addr len -- str-addr len 0 ) ( on error )
( str-addr len -- args... handler-ptr ) ( on success )

The handler pointer is a word which interprets the args, either
compiling or executing depending on the value of the state variable.
This standard interface makes it easy for interpret to check each
in turn. It then calls the handler pointer. Other users of the
words can ignore the exact value of the handler, and treat it as
a flag indicating success or failure.

TODO - refactor ParseInstruction to use the stack instead of
taking args in TMP.
- Allow Forth code to modify the list of handlers to try somehow.